Solution to Simple Calculator

+++ Simple Calculator@@6788dd63490637cfc44e1ba8/code.c
@@ -1,7 +1,7 @@
 #include <stdio.h>
 int main()
 {
-    int i,j;
+    float i,j;
     char op;
     scanf("%d %d",&i,&j);
     scanf("%s",&op);
